pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lanelet2anchors"
version = "0.1.0"
authors = [
  { name="Alexander Naumann", email="alex.code@mail.com" },
  { name="Felix Hertlein", email="hertlein.research@gmail.com" } ,
]
description = "Easy anchor generation for autonomous driving based on Lanelet2 maps"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Free for non-commercial use",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
  'shapely >= 1.8.5, < 2',
  'lanelet2 >= 1.2.1, < 2',
  'networkx >= 2.8.8, < 3',
  'fastdtw >= 0.3.4, < 1',
  'scipy >= 1.10.1, < 2',
]

[project.optional-dependencies]
dev = [
  'nuscenes-devkit >= 1.1.10, < 2',
  'black[jupyter] >= 23.3.0, < 24',
  'mypy >= 1.3.0, <2',
  'isort >= 5.12.0, <6',
  'pre-commit >= 3.3.3, < 4',
  'ipykernel >= 6.23.2, < 7',
]
docs = [
    "mkdocs",
    "mkdocstrings[python]",
    "mkdocs-jupyter",
    "mkdocs-bibtex",
    "jupyter_contrib_nbextensions",
]

[project.urls]
"Homepage" = "https://github.com/FelixHertlein/lanelet2anchors"
"Bug Tracker" = "https://github.com/FelixHertlein/lanelet2anchors/issues"
